IS-2032: Refactored MineMoter and EnhetensMoter into /sider folder
